+++ src/App.js
@@ -1,30 +1,9 @@
 import React, { Fragment, useState, useEffect, useCallback } from "react";
-import useFetch from "./hooks/use-fetch";
-import { URL } from "./store/structure";
 
 const App = () => {
-	const [tableData, setTableData] = useState(null);
-	const requestConfig = {
-		url: URL,
-		requestOptions: {
-			method: "GET",
-			mode: "cors",
-			headers: { "Access-Control-Allow-Origin": "*" },
-		},
-	};
-	const transformData = (data) => {
-		setTableData(data);
-	};
-	const { isLoading, error, sendRequest } = useFetch();
-
-	useEffect(() => {
-		sendRequest(requestConfig, transformData);
-	}, [sendRequest]);
-
 	return (
 		<React.Fragment>
-			{isLoading && <h1>LOADING...</h1>}
-			{!isLoading && tableData && <h1>Hello there you {tableData.table[1].username}!</h1>}
+			<h1>Hello</h1>
 		</React.Fragment>
 	);
 };
